Add picklist hierarchy retrieval functionality
A method was added to retrieve dependent picklist values based on control field values. The changes include an update in the ApplicationFormsService.cls to create the getDependentPicklistValues and getPicklistHierarchy methods. The ApplicationFormsController.cls was updated to allow for these methods to be called externally. The application.js file was also updated to call these new picklist methods and populate picklist values dynamically.