[UAT-2] Update form validation and design for second guardian
The logic for forms concerning the second guardian has been updated to handle cases where only one guardian is present. HTML inputs and the CSS were also adjusted, particularly the Vårdnadshavare 2 input. A check to disable and enable the field is in place depending on the requirement for a second guardian.