[UAT-7] Add required validation and error styling to form inputs
The commit introduces required validation to multiple input fields in the application form and adds error CSS styling for dropdown fields when no selection is made. It also redesigns the form to include help messages for users if required fields are not filled. A new class "no-category-selection" is added which highlights the borders in red when no category is selected. Any subsequent selection will remove the class and the red border.